totally remove kernel_product
