--- doc/python/api/pytorch/KernelProduct.rst
@@ -1,20 +1,0 @@
-Kernel product helper
-----------------------
-
-:mod:`pykeops.torch` - The :doc:`kernel_product <../../kernel-product>` helper:
-
-.. rubric:: Summary
-
-.. currentmodule:: pykeops.torch
-.. autosummary::
-    Kernel
-    kernel_product
-
-.. rubric:: Syntax
-
-
-.. autoclass:: Kernel
-    :members:
-
-.. autofunction:: kernel_product
-
