+++ doc/python/index.rst
@@ -15,8 +15,6 @@ We now fully document the public interface of the :mod:`pykeops` module, which i
 
 3. :doc:`The generic_reduction functions <generic-reduction>`: a **math-friendly** set of helper routines for the Genred module.
 
-4. :doc:`kernel_product <kernel-product>`: a specific syntax with **convenient aliases** for operations in kernel spaces. This module is only compatible with PyTorch.
-
 
 .. toctree::
    :maxdepth: 2
@@ -26,7 +24,6 @@ We now fully document the public interface of the :mod:`pykeops` module, which i
    Genred
    generic-solver
    generic-reduction
-   kernel-product
    sparsity
